@@ -10,23 +10,24 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
-        "ewallet/application": "~1.0",
-        "symfony/console": "4.2.8",
-        "vlucas/phpdotenv": "3.3.3"
+        "php": ">=7.4",
+        "ewallet/application": "1.0",
+        "symfony/console": "5.2.1",
+        "symfony/dependency-injection": "5.2.1",
+        "vlucas/phpdotenv": "5.2.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "2.15.0",
-        "infection/infection": "0.13.1",
-        "nelmio/alice": "3.5.7",
-        "phpstan/phpstan": "0.11.8",
-        "phpstan/phpstan-strict-rules": "0.11.1",
-        "phpunit/phpunit": "8.1.6",
+        "friendsofphp/php-cs-fixer": "2.17.3",
+        "infection/infection": "0.20.2",
+        "nelmio/alice": "3.7.4",
+        "phpstan/phpstan": "0.12.64",
+        "phpstan/phpstan-strict-rules": "0.12.7",
+        "phpunit/phpunit": "9.4.4",
         "roave/security-advisories": "dev-master"
     },
     "config": {
         "platform": {
-            "php": "7.2.19"
+            "php": "7.4.12"
         },
         "sort-packages": true
     },
