@@ -1,21 +1,20 @@
-<?php
+<?php declare(strict_types=1);
 /**
- * PHP version 7.1
+ * PHP version 7.4
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
 
 require __DIR__ . '/../vendor/autoload.php';
 
-use Dotenv\Dotenv;
+use Adapters\Symfony\Application\DependencyInjection\ContainerFactory;
+use Application\BasePath;
+use Application\Environment;
 use Setup\SetupApplication;
 
-if (!isset($_ENV['APP_ENV'])) {
-    // We' re not running the application from the containers
-    $environment = Dotenv::createImmutable(__DIR__ . '/../', '.env');
-    $environment->load();
-    $environment->required(['APP_ENV']);
-}
-
-$application = new SetupApplication(require __DIR__ . '/../setup.php');
-$application->run();
+(static function(): void {
+    $basePath = new BasePath(new SplFileInfo(__DIR__ . '/../'));
+    $environment = Environment::fromGlobals($basePath);
+    $application = SetupApplication::fromContainer(ContainerFactory::new($basePath, $environment));
+    $application->run();
+})();
