@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * PHP version 7.4
  *
@@ -24,31 +24,36 @@ abstract class MembersTest extends TestCase
     /** @test */
     function it_finds_a_registered_member()
     {
-        $registeredMemberId = $this->registeredMember->id();
+        $this->members = $this->membersInstance();
+        $registeredMember = A::member()->build();
+        $this->members->save($registeredMember);
+        $registeredMemberId = $registeredMember->id();
 
         $member = $this->members->with($registeredMemberId);
 
         $this->assertTrue(
-            $member->equals($this->registeredMember),
-            'Registered member with ID "abcd" should be found'
+            $member->equals($registeredMember),
+            "Registered member with ID '{$registeredMemberId}' should be found"
         );
     }
 
     /** @test */
     function it_fails_to_find_an_unknown_member()
     {
-        $this->expectException(UnknownMember::class);
+        $this->members = $this->membersInstance();
 
+        $this->expectException(UnknownMember::class);
         $this->members->with(new MemberId('unknown member id'));
     }
 
     /** @test */
     function it_updates_the_information_of_a_registered_member()
     {
+        $this->members = $this->membersInstance();
         $fivePesos = 500;
         $recipient = A::member()->build();
-        $sender = $this->members->with($this->sender->id());
-
+        $sender = A::member()->withBalance(1000)->build();
+        $this->members->save($sender);
         $sender->transfer(Money::MXN($fivePesos), $recipient);
         $this->members->save($sender);
 
@@ -57,18 +62,6 @@ abstract class MembersTest extends TestCase
             $this->members->with($sender->id()),
             "Current member balance should be $fivePesos"
         );
-    }
-
-    /** @before */
-    function generateFixtures(): void
-    {
-        $this->members = $this->membersInstance();
-        $this->registeredMember = A::member()->build();
-        $this->sender = A::member()->withBalance(1000)->build();
-
-        $this->members->save($this->sender);
-        $this->members->save($this->registeredMember);
-        $this->members->save(A::member()->build());
     }
 
     /** @var Members */
