@@ -1,27 +1,24 @@
 <?php declare(strict_types=1);
 /**
- * PHP version 7.2
+ * PHP version 7.4
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
 
 require __DIR__ . '/../vendor/autoload.php';
 
-use Dotenv\Dotenv;
-use Ewallet\Pimple\EwalletConsoleContainer;
+use Adapters\Symfony\Application\DependencyInjection\ContainerFactory;
+use Application\BasePath;
+use Application\Environment;
 use Ewallet\Ui\Console\EwalletApplication;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-if (! isset($_ENV['APP_ENV'])) {
-    // We' re not running the application from the containers
-    $environment = Dotenv::create(__DIR__ . '/../');
-    $environment->load();
-    $environment->required(['APP_ENV', 'DB_URL']);
-}
+(static function (): void {
+    $basePath = new BasePath(new SplFileInfo(__DIR__ . '/../'));
+    $environment = Environment::fromGlobals($basePath);
+    $container = ContainerFactory::new($basePath, $environment);
+    $application = EwalletApplication::fromContainer($container);
 
-$application = new EwalletApplication($container = new EwalletConsoleContainer(
-    require __DIR__ . '/../config.php'
-));
-
-$application->run($container[InputInterface::class], $container[OutputInterface::class]);
+    $application->run($container->get(InputInterface::class), $container->get(OutputInterface::class));
+})();
