@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 /**
- * PHP version 7.2
+ * PHP version 7.4
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
@@ -15,7 +15,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 /**
  * Interacts with a sender in a console application to transfer funds to a recipient
  */
-class TransferFundsConsole
+final class TransferFundsConsole
 {
     /** @var OutputInterface */
     private $output;
@@ -38,7 +38,7 @@ class TransferFundsConsole
     {
         $this->output->writeln("<comment>[{$errorCode}] Please fix the following errors</comment>");
 
-        array_map(function (string $message) {
+        array_map(function (string $message): void {
             $this->output->writeln("<error>{$message}</error>");
         }, $messages);
 
