@@ -1,24 +1,20 @@
-<?php
+<?php declare(strict_types=1);
 /**
- * PHP version 7.1
+ * PHP version 7.4
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
 
 require __DIR__ . '/../vendor/autoload.php';
 
-use Adapters\Symfony\DependencyInjection\ContainerFactory;
-use Dotenv\Dotenv;
+use Adapters\Symfony\Application\DependencyInjection\ContainerFactory;
+use Application\BasePath;
+use Application\Environment;
 use Framework\Slim\ApplicationFactory;
 
 (static function () {
-    if (!isset($_ENV['APP_ENV'])) {
-        // We' re not running the application from the containers
-        $environment = Dotenv::createImmutable(__DIR__ . '/../');
-        $environment->load();
-        $environment->required(['APP_ENV', 'DB_URL', 'PDO_DRIVER']);
-    }
-
-    $app = ApplicationFactory::createFromContainer(ContainerFactory::new());
+    $basePath = new BasePath(new SplFileInfo(__DIR__ . '/../'));
+    $environment = Environment::fromGlobals($basePath);
+    $app = ApplicationFactory::createFromContainer(ContainerFactory::new($basePath, $environment));
     $app->run();
 })();
