@@ -5,13 +5,20 @@
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
 
+use Application\BasePath;
 use Dotenv\Dotenv;
 use Codeception\Util\Autoload;
 
 (static function () {
     Autoload::addNamespace('Page', __DIR__. '/_support/_pages');
 
-    $environment = Dotenv::createImmutable(__DIR__ . '/../', '.env.tests');
+    $basePath = new BasePath(new SplFileInfo(__DIR__ . '/../'));
+    if (file_exists($basePath->cachePath() . '/container-test.php')) {
+        unlink($basePath->cachePath() . '/container-test.php');
+    }
+    if (file_exists($basePath->cachePath() . '/container-test.php.meta')) {
+        unlink($basePath->cachePath() . '/container-test.php.meta');
+    }
+    $environment = Dotenv::createImmutable($basePath->absolutePath(), '.env.test');
     $environment->load();
-    $environment->required(['APP_ENV', 'DB_URL']);
 })();
