@@ -7,10 +7,17 @@
 
 require __DIR__ . '/../vendor/autoload.php';
 
+use Application\BasePath;
 use Dotenv\Dotenv;
 
 (static function () {
-    $environment = Dotenv::createImmutable(__DIR__ . '/../', '.env.tests');
+    $basePath = new BasePath(new SplFileInfo(__DIR__ . '/../'));
+    if (file_exists($basePath->cachePath() . '/container-test.php')) {
+        unlink($basePath->cachePath() . '/container-test.php');
+    }
+    if (file_exists($basePath->cachePath() . '/container-test.php.meta')) {
+        unlink($basePath->cachePath() . '/container-test.php.meta');
+    }
+    $environment = Dotenv::createImmutable($basePath->absolutePath(), '.env.test');
     $environment->load();
-    $environment->required(['APP_ENV']);
 })();
