@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * PHP version 7.4
  *
@@ -7,8 +7,11 @@
 
 namespace Ewallet\Pimple;
 
-use Adapters\Symfony\DependencyInjection\ContainerFactory;
+use Adapters\Symfony\Application\DependencyInjection\ContainerFactory;
+use Application\BasePath;
+use Application\Environment;
 use PHPUnit\Framework\TestCase;
+use SplFileInfo;
 use Symfony\Component\Finder\Finder;
 
 final class EwalletWebContainerTest extends TestCase
@@ -16,7 +19,8 @@ final class EwalletWebContainerTest extends TestCase
     /** @test */
     function it_creates_the_web_application_controllers()
     {
-        $container = ContainerFactory::new();
+        $basePath = new BasePath(new SplFileInfo(__DIR__ . '/../../../'));
+        $container = ContainerFactory::new($basePath, new Environment('test', true));
         $finder = new Finder();
         $files = $finder->files()->name('*.php')->in(__DIR__ . '/../../../src/UI/Slim/Controllers');
         foreach ($files as $file) {
